Revert "Jsw handler"
