+++ src/app/app.component.html
@@ -1,5 +1,8 @@
+
+
 <app-navbar></app-navbar>
 <app-search-bar></app-search-bar>
-<app-data-cards></app-data-cards>
 
 <router-outlet></router-outlet>
+
+
