--- src/app/elements/components/data-cards/data-cards.component.html
@@ -1,29 +1,0 @@
-
-
-<!--
-<div class="d-flex flex-row justify-content-center align-items-center">
-  <div class="card mb-3" style="max-width: 50rem">
-    <div class="row g-0">
-      <div class="col-md-4">
-        <img
-          src="{{datos.imagen}}"
-          class="img-fluid rounded-start"
-          alt="jaja soy otaku"
-        />
-      </div>
-      <div class="col-md-8">
-        <div class="card-body">
-          <h5 class="card-title">{{datos.titulo}}</h5>
-          <p class="card-text">
-            {{datos.descripcion}}
-          </p>
-          <p class="card-text">
-            <small class="text-muted">Last updated 3 mins ago</small>
-          </p>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
--->
