+++ src/app/elements/data-cards/data-cards.component.html
@@ -1,0 +1,6 @@
+<div class="card" style="width: 18rem;">
+    <img src="https://www.ecured.cu/images/5/50/Bring_Me_The_Horizon.jpg" class="card-img-top" >
+    <div class="card-body">
+      <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
+    </div>
+  </div>
