+++ src/app/elements/data-cards/data-cards.component.ts
@@ -1,4 +1,4 @@
-import { Component,Input } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-data-cards',
@@ -6,5 +6,5 @@ import { Component,Input } from '@angular/core';
   styleUrls: ['./data-cards.component.css']
 })
 export class DataCardsComponent {
-  @Input() datos: any;
+
 }
