+++ src/app/elements/elements.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { SidebarComponent } from './components/sidebar/sidebar.component';
-import { SearchBarComponent } from './components/search-bar/search-bar.component';
-import { DataCardsComponent } from './components/data-cards/data-cards.component';
+import { NavbarComponent } from './navbar/navbar.component';
+import { SidebarComponent } from './sidebar/sidebar.component';
+import { SearchBarComponent } from './search-bar/search-bar.component';
+import { DataCardsComponent } from './data-cards/data-cards.component';
