+++ src/app/elements/search-bar/search-bar.component.css
