+++ src/app/elements/search-bar/search-bar.component.spec.ts
