+++ src/app/logins/login/login.component.html
@@ -1,21 +1,1 @@
-<div class="login">
-  <form name="form" method="post" (ngSubmit)="f.form.valid && onSubmit()" #f="ngForm" novalidate>
-    <h1>Login</h1>
-    <input
-      name="usuario"
-      placeholder="usuario"
-      required="required"
-      [(ngModel)]="form.usuario"
-      #usuario="ngModel"
-    />
-    <input
-      type="password"
-      name="contraseña"
-      placeholder="contraseña"
-      required="required"
-      [(ngModel)]="form.contrasena"
-      #contrasena="ngModel"
-    />
-    <button type="submit">Log in</button>
-  </form>
-</div>
+<p>login works!</p>
