--- src/app/modules/routing/app-routing.module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from '../../home/home.component';
-import { LoginsComponent } from '../../logins/logins.component';
-
-const routes: Routes = [
-  {
-    path: 'logins',
-    component: LoginsComponent,
-  },
-  {
-    path: '',
-    component: HomeComponent,
-  },
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule],
-})
-export class AppRoutingModule {}
